Extract shared tunnel scene; make home.js / worker thin hosts
The three.js tunnel + 2D starfield simulation was duplicated, line for
line, between the main-thread renderer (home.js) and the OffscreenCanvas
worker (tunnelWorker.js). That split is what produced the recurring
"renderer A changed but renderer B didn't" class of bug (e.g. the recent
THREE-null crash).

The simulation now lives once in landing/tunnelScene.js, an
environment-agnostic module that never touches window/document/self.
Everything crossing the environment boundary is injected: the canvases and
viewport come in as parameters; DOM-style output goes out through
onDomUpdate / onShift callbacks. The host owns the rAF loop + scroll
smoothing and calls scene.renderFrame() once per tick.

- home.js is now a main-thread host: it owns the DOM (cards, progress,
  vapor styles), scroll input and boot/splash autoplay, and applies scene
  output via a single applyDomUpdate()/applyShift() pair shared by BOTH the
  worker's messages and the fallback scene's callbacks. The scattered
  `if (!supportsOffscreen)` sim blocks collapse into one async
  initMainThread() (dynamically imports three so it never reaches the main
  bundle when the worker renders).
- tunnelWorker.js is now a thin worker host: transfer canvases in, run the
  loop, forward scene output as messages.

The scene caps the raw devicePixelRatio it is handed using quality.dprCap
(both hosts now pass the raw DPR), so default-quality output is pixel
-identical; only the niche ?dpr override no longer propagates into the
worker, which uses its own cap.
